feat: streamline error handling and validation in Runs service
